Feat: 7주차-folder 페이지의 링크들 가져오기 구현, folder 페이지의 링크 프로퍼티 이름(created_at, image_source)이 sharedPage의 것(createdAt, imageSource)과 달라서 map 함수로 프로퍼티 이름 변경
