+++ src/components/main/Folders.js
@@ -1,0 +1,9 @@
+export function Folders() {
+  return (
+    <>
+      <div>전체.객체의 이름을 프롭으로 받아서 map</div>
+    </>
+  );
+}
+
+export default Folders;
