+++ src/components/pages/SharedPage.js
@@ -12,6 +12,10 @@ function SharedPage() {
     profileImageSource: "",
   });
 
+  const [folderDatas, setFolderDatas] = useState({});
+  const [ownerDatas, setOwnerDatas] = useState({});
+  const [links, setLinks] = useState([]);
+
   const getSampleUserData = async () => {
     try {
       const response = await getData("sample/user");
@@ -28,14 +32,30 @@ function SharedPage() {
       console.log(e);
     }
   };
+  const getLinks = async () => {
+    try {
+      const result = await getData("sample/folder");
+      // console.log("Main", result);
+      const linksData = result.folder.links;
+      const foldersData = result.folder;
+      const ownerData = result.folder.owner;
+
+      setFolderDatas(foldersData);
+      setLinks(linksData);
+      setOwnerDatas(ownerData);
+    } catch (e) {
+      console.log(`getLinks에서 ${e} 오류`);
+    }
+  };
 
   useEffect(() => {
     getSampleUserData();
+    getLinks();
   }, []);
   return (
     <>
       <Header profileDatas={profileDatas} />
-      <Main />
+      <Main ownerDatas={ownerDatas} folderDatas={folderDatas} links={links} />
       <Footer />
     </>
   );
